+++ src/dgisim/helper/quality_of_life.py
@@ -19,6 +19,30 @@ _T = TypeVar('_T')
 BIG_INT = 0x7fffffff
 
 
+class _CachedClassProperty:
+    class _Null(Enum):
+        NULL = 0
+
+    def __init__(self, fget):
+        self._fget = fget
+        self._cache = _CachedClassProperty._Null.NULL
+
+    def __get__(self, obj, klass=None):
+        if self._cache is not _CachedClassProperty._Null.NULL:
+            return self._cache
+        if klass is None:
+            klass = type(obj)
+        self._cache = self._fget.__get__(obj, klass)()
+        return self._cache
+
+
+def cachedclassproperty(func) -> _CachedClassProperty:
+    if not isinstance(func, classmethod):
+        func = classmethod(func)
+
+    return _CachedClassProperty(func)
+
+
 def case_val(condition: bool, first: _T, second: _T) -> _T:
     """ if condition is True; then return first; else return second """
     return first if condition else second