+++ src/dgisim/status/status.py
@@ -30,7 +30,7 @@ from ..effect.structs import StaticTarget, DamageType
 from ..element import Element, Reaction
 from ..event import *
 from ..helper.hashable_dict import HashableDict
-from ..helper.quality_of_life import BIG_INT, case_val, classproperty
+from ..helper.quality_of_life import BIG_INT, cachedclassproperty, case_val
 from .enums import Preprocessables, Informables
 
 if TYPE_CHECKING:
@@ -105,6 +105,7 @@ __all__ = [
     "CrownOfWatatsumiStatus",
     "DeepwoodMemoriesStatus",
     "EchoesOfAnOfferingStatus",
+    "EmblemOfSeveredFateStatus",
     "FlowingRingsStatus",
     "GamblersEarringsStatus",
     "GeneralsAncientHelmStatus",
@@ -744,14 +745,14 @@ class EquipmentStatus(PersonalStatus):
     """
     Basic status, describing weapon, artifact and character unique talents
     """
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.EquipmentCard]:
         raise NotImplementedError()
 
 
 @dataclass(frozen=True)
 class TalentEquipmentStatus(EquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.EquipmentCard]:
         raise NotImplementedError()
 
@@ -760,7 +761,7 @@ class TalentEquipmentStatus(EquipmentStatus):
 class WeaponEquipmentStatus(EquipmentStatus):
     WEAPON_TYPE: ClassVar[WeaponType]
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         raise NotImplementedError()
 
@@ -796,7 +797,7 @@ class WeaponEquipmentStatus(EquipmentStatus):
 
 @dataclass(frozen=True)
 class ArtifactEquipmentStatus(EquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         raise NotImplementedError(cls)
 
@@ -1310,7 +1311,7 @@ class AmosBowStatus(WeaponEquipmentStatus, _UsageLivingStatus):
     activated: bool = False
     ADDITIONAL_DMG_BOOST: ClassVar[int] = 2
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import AmosBow
         return AmosBow
@@ -1379,7 +1380,7 @@ class ElegyForTheEndStatus(WeaponEquipmentStatus):
         TriggeringSignal.POST_SKILL,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import ElegyForTheEnd
         return ElegyForTheEnd
@@ -1420,7 +1421,7 @@ class ElegyForTheEndStatus(WeaponEquipmentStatus):
 class KingsSquireStatus(WeaponEquipmentStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.BOW
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import KingsSquire
         return KingsSquire
@@ -1430,7 +1431,7 @@ class KingsSquireStatus(WeaponEquipmentStatus):
 class RavenBowStatus(WeaponEquipmentStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.BOW
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import RavenBow
         return RavenBow
@@ -1440,7 +1441,7 @@ class RavenBowStatus(WeaponEquipmentStatus):
 class SacrificialBowStatus(_SacrificialWeaponStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.BOW
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import SacrificialBow
         return SacrificialBow
@@ -1459,7 +1460,7 @@ class AThousandFloatingDreamsStatus(WeaponEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import AThousandFloatingDreams
         return AThousandFloatingDreams
@@ -1504,7 +1505,7 @@ class AThousandFloatingDreamsStatus(WeaponEquipmentStatus, _UsageLivingStatus):
 class FruitOfFulfillmentStatus(WeaponEquipmentStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.CATALYST
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import FruitOfFulfillment
         return FruitOfFulfillment
@@ -1514,7 +1515,7 @@ class FruitOfFulfillmentStatus(WeaponEquipmentStatus):
 class MagicGuideStatus(WeaponEquipmentStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.CATALYST
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import MagicGuide
         return MagicGuide
@@ -1524,7 +1525,7 @@ class MagicGuideStatus(WeaponEquipmentStatus):
 class SacrificialFragmentsStatus(_SacrificialWeaponStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.CATALYST
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import SacrificialFragments
         return SacrificialFragments
@@ -1536,7 +1537,7 @@ class SacrificialFragmentsStatus(_SacrificialWeaponStatus):
 class SacrificialGreatswordStatus(_SacrificialWeaponStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.CLAYMORE
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import SacrificialGreatsword
         return SacrificialGreatsword
@@ -1554,7 +1555,7 @@ class TheBellStatus(WeaponEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import TheBell
         return TheBell
@@ -1599,7 +1600,7 @@ class TheBellStatus(WeaponEquipmentStatus, _UsageLivingStatus):
 class WhiteIronGreatswordStatus(WeaponEquipmentStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.CLAYMORE
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import WhiteIronGreatsword
         return WhiteIronGreatsword
@@ -1611,7 +1612,7 @@ class WolfsGravestoneStatus(WeaponEquipmentStatus):
     HP_THRESHOLD: ClassVar[int] = 6
     ADDITIONAL_DMG_BOOST: ClassVar[int] = 2
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import WolfsGravestone
         return WolfsGravestone
@@ -1645,7 +1646,7 @@ class EngulfingLightningStatus(WeaponEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import EngulfingLightning
         return EngulfingLightning
@@ -1668,7 +1669,7 @@ class EngulfingLightningStatus(WeaponEquipmentStatus, _UsageLivingStatus):
                 return [
                     eft.EnergyRechargeEffect(
                         target=source,
-                        recharge=1,
+                        amount=1,
                     ),
                 ], replace(self, usages=-1)
         elif signal is TriggeringSignal.ROUND_END and self.usages < self.MAX_USAGES:
@@ -1680,7 +1681,7 @@ class EngulfingLightningStatus(WeaponEquipmentStatus, _UsageLivingStatus):
 class LithicSpearStatus(WeaponEquipmentStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.POLEARM
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import LithicSpear
         return LithicSpear
@@ -1690,7 +1691,7 @@ class LithicSpearStatus(WeaponEquipmentStatus):
 class MoonpiercerStatus(WeaponEquipmentStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.POLEARM
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import Moonpiercer
         return Moonpiercer
@@ -1701,7 +1702,7 @@ class VortexVanquisherStatus(WeaponEquipmentStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.POLEARM
     ADDITIONAL_DMG_BOOST: ClassVar[int] = 1
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import VortexVanquisher
         return VortexVanquisher
@@ -1734,7 +1735,7 @@ class VortexVanquisherStatus(WeaponEquipmentStatus):
 class WhiteTasselStatus(WeaponEquipmentStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.POLEARM
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import WhiteTassel
         return WhiteTassel
@@ -1755,7 +1756,7 @@ class AquilaFavoniaStatus(WeaponEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import AquilaFavonia
         return AquilaFavonia
@@ -1811,7 +1812,7 @@ class FavoniusSwordStatus(WeaponEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import FavoniusSword
         return FavoniusSword
@@ -1847,7 +1848,7 @@ class FavoniusSwordStatus(WeaponEquipmentStatus, _UsageLivingStatus):
                 return [
                     eft.EnergyRechargeEffect(
                         target=source,
-                        recharge=self.ENERGY_RECHARGE_AMOUNT,
+                        amount=self.ENERGY_RECHARGE_AMOUNT,
                     ),
                 ], replace(self, usages=-1, activated=False)
             else:
@@ -1861,7 +1862,7 @@ class FavoniusSwordStatus(WeaponEquipmentStatus, _UsageLivingStatus):
 class SacrificialSwordStatus(_SacrificialWeaponStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.SWORD
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import SacrificialSword
         return SacrificialSword
@@ -1871,7 +1872,7 @@ class SacrificialSwordStatus(_SacrificialWeaponStatus):
 class TravelersHandySwordStatus(WeaponEquipmentStatus):
     WEAPON_TYPE: ClassVar[WeaponType] = WeaponType.SWORD
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.WeaponEquipmentCard]:
         from ..card.card import TravelersHandySword
         return TravelersHandySword
@@ -1955,7 +1956,7 @@ class _ElementalDiscountSupplyStatus(_ElementalDiscountStatus):
 class ArchaicPetraStatus(_ElementalDiscountSupplyStatus):
     _ELEMENT: ClassVar[Element] = Element.GEO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import ArchaicPetra
         return ArchaicPetra
@@ -1964,7 +1965,7 @@ class ArchaicPetraStatus(_ElementalDiscountSupplyStatus):
 class BlizzardStrayerStatus(_ElementalDiscountSupplyStatus):
     _ELEMENT: ClassVar[Element] = Element.CRYO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import BlizzardStrayer
         return BlizzardStrayer
@@ -1973,7 +1974,7 @@ class BlizzardStrayerStatus(_ElementalDiscountSupplyStatus):
 class BrokenRimesEchoStatus(_ElementalDiscountStatus):
     _ELEMENT: ClassVar[Element] = Element.CRYO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import BrokenRimesEcho
         return BrokenRimesEcho
@@ -1983,7 +1984,7 @@ class BrokenRimesEchoStatus(_ElementalDiscountStatus):
 class CrimsonWitchOfFlamesStatus(_ElementalDiscountSupplyStatus):
     _ELEMENT: ClassVar[Element] = Element.PYRO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import CrimsonWitchOfFlames
         return CrimsonWitchOfFlames
@@ -1999,7 +2000,7 @@ class CrownOfWatatsumiStatus(ArtifactEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.POST_HEALING,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import CrownOfWatatsumi
         return CrownOfWatatsumi
@@ -2049,7 +2050,7 @@ class CrownOfWatatsumiStatus(ArtifactEquipmentStatus, _UsageLivingStatus):
 class DeepwoodMemoriesStatus(_ElementalDiscountSupplyStatus):
     _ELEMENT: ClassVar[Element] = Element.DENDRO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import DeepwoodMemories
         return DeepwoodMemories
@@ -2064,7 +2065,7 @@ class EchoesOfAnOfferingStatus(ArtifactEquipmentStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import EchoesOfAnOffering
         return EchoesOfAnOffering
@@ -2105,6 +2106,51 @@ class EchoesOfAnOfferingStatus(ArtifactEquipmentStatus):
 
 
 @dataclass(frozen=True, kw_only=True)
+class EmblemOfSeveredFateStatus(ArtifactEquipmentStatus):
+    DMG_BOOST: ClassVar[int] = 2
+    REACTABLE_SIGNALS: ClassVar[frozenset[TriggeringSignal]] = frozenset((
+        TriggeringSignal.POST_SKILL,
+    ))
+
+    @cachedclassproperty
+    def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
+        from ..card.card import EmblemOfSeveredFate
+        return EmblemOfSeveredFate
+
+    @override
+    def _preprocess(
+            self, game_state: GameState, status_source: StaticTarget, item: PreprocessableEvent,
+            signal: Preprocessables,
+    ) -> tuple[PreprocessableEvent, None | Self]:
+        if signal is Preprocessables.DMG_AMOUNT_PLUS:
+            assert isinstance(item, DmgPEvent)
+            if (
+                    item.dmg.source == status_source
+                    and item.dmg.damage_type.direct_elemental_burst()
+            ):
+                return item.delta_damage(self.DMG_BOOST), self
+        return item, self
+
+    @override
+    def _react_to_signal(
+            self, game_state: GameState, source: StaticTarget, signal: TriggeringSignal,
+            detail: None | InformableEvent
+    ) -> tuple[list[eft.Effect], None | Self]:
+        if signal is TriggeringSignal.POST_SKILL:
+            assert isinstance(detail, SkillIEvent)
+            if (
+                    detail.source.pid is source.pid
+                    and detail.source.id != source.id
+                    and detail.skill_type.is_elemental_burst()
+            ):
+                return [eft.EnergyRechargeEffect(
+                    target=source,
+                    amount=1,
+                )], self
+        return [], self
+
+
+@dataclass(frozen=True, kw_only=True)
 class FlowingRingsStatus(ArtifactEquipmentStatus, _UsageLivingStatus):
     usages: int = 1
     activated: bool = False
@@ -2113,7 +2159,7 @@ class FlowingRingsStatus(ArtifactEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import FlowingRings
         return FlowingRings
@@ -2160,7 +2206,7 @@ class GamblersEarringsStatus(ArtifactEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.POST_DMG,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import GamblersEarrings
         return GamblersEarrings
@@ -2195,7 +2241,7 @@ class GeneralsAncientHelmStatus(ArtifactEquipmentStatus):
         TriggeringSignal.ROUND_START,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import GeneralsAncientHelm
         return GeneralsAncientHelm
@@ -2248,7 +2294,7 @@ class GildedDreamsStatus(_ShadowOfTheSandKingLikeStatus):
     usages: int = 2
     MAX_USAGES: ClassVar[int] = 2
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import GildedDreams
         return GildedDreams
@@ -2288,7 +2334,7 @@ class _HeartOfKhvarenasBrillianceLikeStatus(ArtifactEquipmentStatus, _UsageLivin
 class HeartOfDepthStatus(_ElementalDiscountSupplyStatus):
     _ELEMENT: ClassVar[Element] = Element.HYDRO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import HeartOfDepth
         return HeartOfDepth
@@ -2296,7 +2342,7 @@ class HeartOfDepthStatus(_ElementalDiscountSupplyStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class HeartOfKhvarenasBrillianceStatus(_HeartOfKhvarenasBrillianceLikeStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import HeartOfKhvarenasBrilliance
         return HeartOfKhvarenasBrilliance
@@ -2312,7 +2358,7 @@ class InstructorsCapStatus(ArtifactEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import InstructorsCap
         return InstructorsCap
@@ -2361,7 +2407,7 @@ class InstructorsCapStatus(ArtifactEquipmentStatus, _UsageLivingStatus):
 class LaurelCoronetStatus(_ElementalDiscountStatus):
     _ELEMENT: ClassVar[Element] = Element.DENDRO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import LaurelCoronet
         return LaurelCoronet
@@ -2371,7 +2417,7 @@ class LaurelCoronetStatus(_ElementalDiscountStatus):
 class MaskOfSolitudeBasaltStatus(_ElementalDiscountStatus):
     _ELEMENT: ClassVar[Element] = Element.GEO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import MaskOfSolitudeBasalt
         return MaskOfSolitudeBasalt
@@ -2382,7 +2428,7 @@ class ShadowOfTheSandKingStatus(_ShadowOfTheSandKingLikeStatus):
     usages: int = 1
     MAX_USAGES: ClassVar[int] = 1
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import ShadowOfTheSandKing
         return ShadowOfTheSandKing
@@ -2399,7 +2445,7 @@ class TenacityOfTheMillelithStatus(ArtifactEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import TenacityOfTheMillelith
         return TenacityOfTheMillelith
@@ -2443,7 +2489,7 @@ class TenacityOfTheMillelithStatus(ArtifactEquipmentStatus, _UsageLivingStatus):
 class ThunderSummonersCrownStatus(_ElementalDiscountStatus):
     _ELEMENT: ClassVar[Element] = Element.ELECTRO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import ThunderSummonersCrown
         return ThunderSummonersCrown
@@ -2453,7 +2499,7 @@ class ThunderSummonersCrownStatus(_ElementalDiscountStatus):
 class ThunderingFuryStatus(_ElementalDiscountSupplyStatus):
     _ELEMENT: ClassVar[Element] = Element.ELECTRO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import ThunderingFury
         return ThunderingFury
@@ -2463,7 +2509,7 @@ class ThunderingFuryStatus(_ElementalDiscountSupplyStatus):
 class ViridescentVenererStatus(_ElementalDiscountSupplyStatus):
     _ELEMENT: ClassVar[Element] = Element.ANEMO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import ViridescentVenerer
         return ViridescentVenerer
@@ -2473,7 +2519,7 @@ class ViridescentVenererStatus(_ElementalDiscountSupplyStatus):
 class ViridescentVenerersDiademStatus(_ElementalDiscountStatus):
     _ELEMENT: ClassVar[Element] = Element.ANEMO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import ViridescentVenerersDiadem
         return ViridescentVenerersDiadem
@@ -2486,7 +2532,7 @@ class VourukashasGlowStatus(_HeartOfKhvarenasBrillianceLikeStatus):
         TriggeringSignal.END_ROUND_CHECK_OUT,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import VourukashasGlow
         return VourukashasGlow
@@ -2511,7 +2557,7 @@ class VourukashasGlowStatus(_HeartOfKhvarenasBrillianceLikeStatus):
 class WineStainedTricorneStatus(_ElementalDiscountStatus):
     _ELEMENT: ClassVar[Element] = Element.HYDRO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import WineStainedTricorne
         return WineStainedTricorne
@@ -2521,7 +2567,7 @@ class WineStainedTricorneStatus(_ElementalDiscountStatus):
 class WitchsScorchingHatStatus(_ElementalDiscountStatus):
     _ELEMENT: ClassVar[Element] = Element.PYRO
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.ArtifactEquipmentCard]:
         from ..card.card import WitchsScorchingHat
         return WitchsScorchingHat
@@ -2538,10 +2584,7 @@ class AncientCourtyardStatus(CombatStatus):
 
     @override
     def _preprocess(
-            self,
-            game_state: GameState,
-            status_source: StaticTarget,
-            item: PreprocessableEvent,
+            self, game_state: GameState, status_source: StaticTarget, item: PreprocessableEvent,
             signal: Preprocessables,
     ) -> tuple[PreprocessableEvent, None | Self]:
         if signal is Preprocessables.CARD1_COST_OMNI:
@@ -3655,7 +3698,7 @@ Group statues by characters, characters ordered alphabetically
 
 @dataclass(frozen=True, kw_only=True)
 class DescentOfDivinityStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import DescentOfDivinity
         return DescentOfDivinity
@@ -3673,7 +3716,7 @@ class AratakiIchibanStatus(TalentEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import AratakiIchiban
         return AratakiIchiban
@@ -3821,7 +3864,7 @@ class SuperlativeSuperstrengthStatus(CharacterStatus, _UsageStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class GrandExpectationStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import GrandExpectation
         return GrandExpectation
@@ -3979,7 +4022,7 @@ class ChonghuasFrostFieldStatus(CombatStatus, _InfusionStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class SteadyBreathingStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import SteadyBreathing
         return SteadyBreathing
@@ -4025,7 +4068,7 @@ class ColleiTalentStatus(CharacterHiddenStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class FloralSidewinderStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import FloralSidewinder
         return FloralSidewinder
@@ -4126,7 +4169,7 @@ class StalwartAndTrueStatus(TalentEquipmentStatus):
         TriggeringSignal.END_ROUND_CHECK_OUT,
     })
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import StalwartAndTrue
         return StalwartAndTrue
@@ -4168,7 +4211,7 @@ class CatClawShieldStatus(StackedShieldStatus, CombatStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class ShakenNotPurredStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import ShakenNotPurred
         return ShakenNotPurred
@@ -4353,7 +4396,7 @@ class GrimheartStatus(CharacterStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class WellspingOfWarLustStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import WellspingOfWarLust
         return WellspingOfWarLust
@@ -4363,7 +4406,7 @@ class WellspingOfWarLustStatus(TalentEquipmentStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class CicinsColdGlareStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import CicinsColdGlare
         return CicinsColdGlare
@@ -4379,7 +4422,7 @@ class FlowingCicinShieldStatus(CharacterStatus, StackedShieldStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class PaidInFullStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import PaidInFull
         return PaidInFull
@@ -4449,7 +4492,7 @@ class StealthStatus(CharacterStatus, FixedShieldStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class StellarPredatorStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import StellarPredator
         return StellarPredator
@@ -4489,7 +4532,7 @@ class IceLotusStatus(CombatStatus, FixedShieldStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class UndividedHeartStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import UndividedHeart
         return UndividedHeart
@@ -4532,7 +4575,7 @@ class ParamitaPapilioStatus(CharacterStatus, _InfusionStatus):
 @dataclass(frozen=True, kw_only=True)
 class SanguineRougeStatus(TalentEquipmentStatus):
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import SanguineRouge
         return SanguineRouge
@@ -4563,7 +4606,7 @@ class SanguineRougeStatus(TalentEquipmentStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class ProliferatingSporesStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import ProliferatingSpores
         return ProliferatingSpores
@@ -4692,7 +4735,7 @@ class RadicalVitalityStatus(CharacterStatus, _UsageLivingStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class LandsOfDandelionStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import LandsOfDandelion
         return LandsOfDandelion
@@ -4800,7 +4843,7 @@ _MIDARE_RANZAN_MAP: dict[Element, type[MidareRanzanStatus]] = HashableDict({
 
 @dataclass(frozen=True, kw_only=True)
 class PoeticsOfFuubutsuStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import PoeticsOfFuubutsu
         return PoeticsOfFuubutsu
@@ -4908,7 +4951,7 @@ class ColdBloodedStrikeStatus(TalentEquipmentStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import ColdBloodedStrike
         return ColdBloodedStrike
@@ -5027,7 +5070,7 @@ class KantenSenmyouBlessingStatus(TalentEquipmentStatus, _UsageStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import KantenSenmyouBlessing
         return KantenSenmyouBlessing
@@ -5087,7 +5130,7 @@ class KeqingTalentStatus(CharacterHiddenStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class ThunderingPenanceStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import ThunderingPenance
         return ThunderingPenance
@@ -5149,7 +5192,7 @@ class ExplosiveSparkStatus(CharacterStatus, _UsageStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class PoundingSurpriseStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import PoundingSurprise
         return PoundingSurprise
@@ -5247,7 +5290,7 @@ class CrowfeatherCoverStatus(CharacterStatus, _UsageStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class SinOfPrideStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import SinOfPride
         return SinOfPride
@@ -5263,7 +5306,7 @@ class CurtainOfSlumberStatus(CombatStatus, StackedShieldStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class LightsRemitStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import LightsRemit
         return LightsRemit
@@ -5360,7 +5403,7 @@ class PulsatingWitchStatus(TalentEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import PulsatingWitch
         return PulsatingWitch
@@ -5394,7 +5437,7 @@ class ConclusiveOvationStatus(TalentEquipmentStatus, _UsageLivingStatus):
         TriggeringSignal.ROUND_END,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import ConclusiveOvation
         return ConclusiveOvation
@@ -5479,7 +5522,7 @@ class PropSurplusStatus(CharacterStatus, _UsageLivingStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class TranscendentAutomatonStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import TranscendentAutomaton
         return TranscendentAutomaton
@@ -5549,7 +5592,7 @@ class IllusoryTorrentStatus(CharacterHiddenStatus):
 class ProphecyOfSubmersionStatus(TalentEquipmentStatus):
     DMG_BOOST: ClassVar[int] = 2
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import ProphecyOfSubmersion
         return ProphecyOfSubmersion
@@ -5699,7 +5742,7 @@ class ShrineOfMayaStatus(CombatStatus, _UsageStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class TheSeedOfStoredKnowledgeStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import TheSeedOfStoredKnowledge
         return TheSeedOfStoredKnowledge
@@ -5747,7 +5790,7 @@ class JadeScreenStatus(CombatStatus, FixedShieldStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class StrategicReserveStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import StrategicReserve
         return StrategicReserve
@@ -5814,7 +5857,7 @@ class FullPlateStatus(CombatStatus, StackedShieldStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class IGotYourBackStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import IGotYourBack
         return IGotYourBack
@@ -5947,7 +5990,7 @@ class QiqiTalentStatus(CharacterHiddenStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class RiteOfResurrectionStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import RiteOfResurrection
         return RiteOfResurrection
@@ -6028,7 +6071,7 @@ class ChakraDesiderataStatus(CharacterStatus, _UsageLivingStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class WishesUnnumberedStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import WishesUnnumbered
         return WishesUnnumbered
@@ -6039,7 +6082,7 @@ class WishesUnnumberedStatus(TalentEquipmentStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class StreamingSurgeStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import StreamingSurge
         return StreamingSurge
@@ -6099,7 +6142,7 @@ class CeremonialGarmentStatus(CharacterStatus, _UsageStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class TamakushiCasketStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import TamakushiCasket
         return TamakushiCasket
@@ -6177,7 +6220,7 @@ class IcyQuillStatus(CombatStatus, _UsageStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class MysticalAbandonStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import MysticalAbandon
         return MysticalAbandon
@@ -6237,7 +6280,7 @@ class StonehideReforgedStatus(TalentEquipmentStatus):
         TriggeringSignal.POST_DMG,
     ))
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import StonehideReforged
         return StonehideReforged
@@ -6273,7 +6316,7 @@ class StonehideStatus(CharacterStatus, FixedShieldStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class AbyssalMayhemHydrospoutStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import AbyssalMayhemHydrospout
         return AbyssalMayhemHydrospout
@@ -6464,7 +6507,7 @@ class TideWithholderStatus(CharacterHiddenStatus):
 class KeenSightStatus(TalentEquipmentStatus):
     COST_DEDUCTION: ClassVar[int] = 1
 
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import KeenSight
         return KeenSight
@@ -6579,7 +6622,7 @@ class VijnanaSuffusionStatus(CharacterStatus, _UsageStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class EmbraceOfWindsStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import EmbraceOfWinds
         return EmbraceOfWinds
@@ -6712,7 +6755,7 @@ class DescentStatus(CharacterStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class GalesOfReverieStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import GalesOfReverie
         return GalesOfReverie
@@ -6867,7 +6910,7 @@ class RainbowBladeworkStatus(CombatStatus, _UsageStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class TheScentRemainedStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import TheScentRemained
         return TheScentRemained
@@ -6943,7 +6986,7 @@ class TenkoThunderboltsStatus(CombatStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class TheShrinesSacredShadeStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import TheShrinesSacredShade
         return TheShrinesSacredShade
@@ -6984,7 +7027,7 @@ class AdeptalLegacyStatus(CombatStatus, _UsageStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class BeneficentStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import Beneficent
         return Beneficent
@@ -7097,7 +7140,7 @@ class ExquisiteThrowStatus(CombatStatus, _UsageStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class TurnControlStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import TurnControl
         return TurnControl
@@ -7204,7 +7247,7 @@ class AurousBlazeStatus(CombatStatus, _UsageStatus):
 
 @dataclass(frozen=True, kw_only=True)
 class NaganoharaMeteorSwarmStatus(TalentEquipmentStatus):
-    @classproperty
+    @cachedclassproperty
     def CARD(cls) -> type[crd.TalentEquipmentCard]:
         from ..card.card import NaganoharaMeteorSwarm
         return NaganoharaMeteorSwarm